Add naming and file length hygiene checks
