+++ systemDesign.md
@@ -339,6 +339,7 @@ The test suite remains the primary regression net for:
 - GitHub Actions runs validation with `npm ci` on Node `20.19.0`, `22.13.0`, and `24.x`.
 - `SECURITY.md` owns vulnerability reporting expectations and credential-handling guidance.
 - Repository hygiene tests reject high-confidence secret patterns plus common local environment, archive, log, editor, and temporary files.
+- Repository hygiene tests also enforce conventional file names, lower camelCase function declarations, and maintained-file line-count limits.
 
 ## 11. Repository rules that should remain true
